Add a winner function and more display functions
'winner' will return either Just Black or Just White if black or white
won, otherwise, it will return Nothing to indicate that neither player
won (which, if the function is ran at the end of the game would mean
that it is a draw).

The other functions are there to help print to the user, by converting a
colour to a displayable string.